feat: Make `AnyIdentifiable` public
